Added tests for session expiration
-  tests for new methods on the session model
-  an integration test for expiring a session on the backend
-  selenium test for expiring a session on the frontend

This required modifying test_utils a bit to allow config to be set when
starting servers - in order to set the expire time to a manageably low value.

This has inevitably added to the run time for the selenium tests. The
lowest expire time that safely avoided accidental expiration seems to
be around 30 seconds, so there's a 30 second sleep right at the end.